Refactor to include functions for select, textarea, input
This makes the code easier to understand. We loop through all
the elements of formData, a swtich statement calls the function
that translates the object into a DOM node, then we add that
DOM node to the parent div for the form.

Each of the helper functions takes care of translating the
object into a DOM node.